Convert libpgport's pqsignal() to a void function.
The protections added by commit 3b00fdba9f introduced race
conditions to this function that can lead to bogus return values.
Since nobody seems to inspect the return value, this is of little
consequence, but it would have been nice to convert it to a void
function to avoid any possibility of a bogus return value.
Unfortunately, doing so would have required also modifying
legacy-pqsignal.c's version of the function, which would've
required an SONAME bump.  Or so I thought...

Thanks to commit 9a45a89c38, legacy-pqsignal.c now has its own
dedicated extern for pqsignal(), which decouples it enough that we
can follow through with changing libpgport's pqsignal() to a void
function.  This commit also adds a bit of error checking in the
form of assertions for the return value of sigaction()/signal().
Since a failure most likely indicates a coding error, and nobody
has ever bothered to check pqsignal()'s return value, it's probably
not worth doing anything fancier.